Bid rides API/myproject/requirements.txt
argcomplete==3.0.8
asgiref==3.4.1
boto3==1.26.148
botocore==1.29.148
certifi==2023.5.7
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cx-Oracle==8.3.0
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2021.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
urllib3==1.26.16
Werkzeug==2.3.1
zappa==0.57.0
